+++ src/Aws/OpsWorks/OpsWorksClient.php
@@ -27,6 +27,7 @@ use Guzzle\Service\Resource\ResourceIteratorInterface;
 /**
  * Client to interact with AWS OpsWorks
  *
+ * @method Model assignInstance(array $args = array()) {@command OpsWorks AssignInstance}
  * @method Model assignVolume(array $args = array()) {@command OpsWorks AssignVolume}
  * @method Model associateElasticIp(array $args = array()) {@command OpsWorks AssociateElasticIp}
  * @method Model attachElasticLoadBalancer(array $args = array()) {@command OpsWorks AttachElasticLoadBalancer}
@@ -43,6 +44,7 @@ use Guzzle\Service\Resource\ResourceIteratorInterface;
  * @method Model deleteStack(array $args = array()) {@command OpsWorks DeleteStack}
  * @method Model deleteUserProfile(array $args = array()) {@command OpsWorks DeleteUserProfile}
  * @method Model deregisterElasticIp(array $args = array()) {@command OpsWorks DeregisterElasticIp}
+ * @method Model deregisterInstance(array $args = array()) {@command OpsWorks DeregisterInstance}
  * @method Model deregisterRdsDbInstance(array $args = array()) {@command OpsWorks DeregisterRdsDbInstance}
  * @method Model deregisterVolume(array $args = array()) {@command OpsWorks DeregisterVolume}
  * @method Model describeApps(array $args = array()) {@command OpsWorks DescribeApps}
@@ -58,6 +60,7 @@ use Guzzle\Service\Resource\ResourceIteratorInterface;
  * @method Model describeRaidArrays(array $args = array()) {@command OpsWorks DescribeRaidArrays}
  * @method Model describeRdsDbInstances(array $args = array()) {@command OpsWorks DescribeRdsDbInstances}
  * @method Model describeServiceErrors(array $args = array()) {@command OpsWorks DescribeServiceErrors}
+ * @method Model describeStackProvisioningParameters(array $args = array()) {@command OpsWorks DescribeStackProvisioningParameters}
  * @method Model describeStackSummary(array $args = array()) {@command OpsWorks DescribeStackSummary}
  * @method Model describeStacks(array $args = array()) {@command OpsWorks DescribeStacks}
  * @method Model describeTimeBasedAutoScaling(array $args = array()) {@command OpsWorks DescribeTimeBasedAutoScaling}
@@ -68,6 +71,7 @@ use Guzzle\Service\Resource\ResourceIteratorInterface;
  * @method Model getHostnameSuggestion(array $args = array()) {@command OpsWorks GetHostnameSuggestion}
  * @method Model rebootInstance(array $args = array()) {@command OpsWorks RebootInstance}
  * @method Model registerElasticIp(array $args = array()) {@command OpsWorks RegisterElasticIp}
+ * @method Model registerInstance(array $args = array()) {@command OpsWorks RegisterInstance}
  * @method Model registerRdsDbInstance(array $args = array()) {@command OpsWorks RegisterRdsDbInstance}
  * @method Model registerVolume(array $args = array()) {@command OpsWorks RegisterVolume}
  * @method Model setLoadBasedAutoScaling(array $args = array()) {@command OpsWorks SetLoadBasedAutoScaling}
@@ -77,6 +81,7 @@ use Guzzle\Service\Resource\ResourceIteratorInterface;
  * @method Model startStack(array $args = array()) {@command OpsWorks StartStack}
  * @method Model stopInstance(array $args = array()) {@command OpsWorks StopInstance}
  * @method Model stopStack(array $args = array()) {@command OpsWorks StopStack}
+ * @method Model unassignInstance(array $args = array()) {@command OpsWorks UnassignInstance}
  * @method Model unassignVolume(array $args = array()) {@command OpsWorks UnassignVolume}
  * @method Model updateApp(array $args = array()) {@command OpsWorks UpdateApp}
  * @method Model updateElasticIp(array $args = array()) {@command OpsWorks UpdateElasticIp}