+++ src/Aws/OpsWorks/Resources/opsworks-2013-02-18.php
@@ -31,6 +31,53 @@ return array (
         ),
     ),
     'operations' => array(
+        'AssignInstance' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'OpsWorks_20130218.AssignInstance',
+                ),
+                'InstanceId' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+                'LayerIds' => array(
+                    'required' => true,
+                    'type' => 'array',
+                    'location' => 'json',
+                    'items' => array(
+                        'name' => 'String',
+                        'type' => 'string',
+                    ),
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'Indicates that a request was invalid.',
+                    'class' => 'ValidationException',
+                ),
+                array(
+                    'reason' => 'Indicates that a resource was not found.',
+                    'class' => 'ResourceNotFoundException',
+                ),
+            ),
+        ),
         'AssignVolume' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -840,6 +887,24 @@ return array (
                     'format' => 'boolean-string',
                     'location' => 'json',
                 ),
+                'LifecycleEventConfiguration' => array(
+                    'type' => 'object',
+                    'location' => 'json',
+                    'properties' => array(
+                        'Shutdown' => array(
+                            'type' => 'object',
+                            'properties' => array(
+                                'ExecutionTimeout' => array(
+                                    'type' => 'numeric',
+                                ),
+                                'DelayUntilElbConnectionsDrained' => array(
+                                    'type' => 'boolean',
+                                    'format' => 'boolean-string',
+                                ),
+                            ),
+                        ),
+                    ),
+                ),
             ),
             'errorResponses' => array(
                 array(
@@ -1271,6 +1336,44 @@ return array (
                     'default' => 'OpsWorks_20130218.DeregisterElasticIp',
                 ),
                 'ElasticIp' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'Indicates that a request was invalid.',
+                    'class' => 'ValidationException',
+                ),
+                array(
+                    'reason' => 'Indicates that a resource was not found.',
+                    'class' => 'ResourceNotFoundException',
+                ),
+            ),
+        ),
+        'DeregisterInstance' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'OpsWorks_20130218.DeregisterInstance',
+                ),
+                'InstanceId' => array(
                     'required' => true,
                     'type' => 'string',
                     'location' => 'json',
@@ -1944,6 +2047,44 @@ return array (
                 ),
             ),
         ),
+        'DescribeStackProvisioningParameters' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'DescribeStackProvisioningParametersResult',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'OpsWorks_20130218.DescribeStackProvisioningParameters',
+                ),
+                'StackId' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'Indicates that a request was invalid.',
+                    'class' => 'ValidationException',
+                ),
+                array(
+                    'reason' => 'Indicates that a resource was not found.',
+                    'class' => 'ResourceNotFoundException',
+                ),
+            ),
+        ),
         'DescribeStackSummary' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -2342,6 +2483,76 @@ return array (
                     'required' => true,
                     'type' => 'string',
                     'location' => 'json',
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'Indicates that a request was invalid.',
+                    'class' => 'ValidationException',
+                ),
+                array(
+                    'reason' => 'Indicates that a resource was not found.',
+                    'class' => 'ResourceNotFoundException',
+                ),
+            ),
+        ),
+        'RegisterInstance' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'RegisterInstanceResult',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'OpsWorks_20130218.RegisterInstance',
+                ),
+                'StackId' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+                'Hostname' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+                'PublicIp' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+                'PrivateIp' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+                'RsaPublicKey' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+                'RsaPublicKeyFingerprint' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+                'InstanceIdentity' => array(
+                    'type' => 'object',
+                    'location' => 'json',
+                    'properties' => array(
+                        'Document' => array(
+                            'type' => 'string',
+                        ),
+                        'Signature' => array(
+                            'type' => 'string',
+                        ),
+                    ),
                 ),
             ),
             'errorResponses' => array(
@@ -2865,6 +3076,44 @@ return array (
                 ),
             ),
         ),
+        'UnassignInstance' => array(
+            'httpMethod' => 'POST',
+            'uri' => '/',
+            'class' => 'Aws\\Common\\Command\\JsonCommand',
+            'responseClass' => 'EmptyOutput',
+            'responseType' => 'model',
+            'parameters' => array(
+                'Content-Type' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'application/x-amz-json-1.1',
+                ),
+                'command.expects' => array(
+                    'static' => true,
+                    'default' => 'application/json',
+                ),
+                'X-Amz-Target' => array(
+                    'static' => true,
+                    'location' => 'header',
+                    'default' => 'OpsWorks_20130218.UnassignInstance',
+                ),
+                'InstanceId' => array(
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+            ),
+            'errorResponses' => array(
+                array(
+                    'reason' => 'Indicates that a request was invalid.',
+                    'class' => 'ValidationException',
+                ),
+                array(
+                    'reason' => 'Indicates that a resource was not found.',
+                    'class' => 'ResourceNotFoundException',
+                ),
+            ),
+        ),
         'UnassignVolume' => array(
             'httpMethod' => 'POST',
             'uri' => '/',
@@ -3344,6 +3593,24 @@ return array (
                     'type' => 'boolean',
                     'format' => 'boolean-string',
                     'location' => 'json',
+                ),
+                'LifecycleEventConfiguration' => array(
+                    'type' => 'object',
+                    'location' => 'json',
+                    'properties' => array(
+                        'Shutdown' => array(
+                            'type' => 'object',
+                            'properties' => array(
+                                'ExecutionTimeout' => array(
+                                    'type' => 'numeric',
+                                ),
+                                'DelayUntilElbConnectionsDrained' => array(
+                                    'type' => 'boolean',
+                                    'format' => 'boolean-string',
+                                ),
+                            ),
+                        ),
+                    ),
                 ),
             ),
             'errorResponses' => array(
@@ -4194,6 +4461,26 @@ return array (
                             'EbsOptimized' => array(
                                 'type' => 'boolean',
                             ),
+                            'ReportedOs' => array(
+                                'type' => 'object',
+                                'properties' => array(
+                                    'Family' => array(
+                                        'type' => 'string',
+                                    ),
+                                    'Name' => array(
+                                        'type' => 'string',
+                                    ),
+                                    'Version' => array(
+                                        'type' => 'string',
+                                    ),
+                                ),
+                            ),
+                            'InfrastructureClass' => array(
+                                'type' => 'string',
+                            ),
+                            'RegisteredBy' => array(
+                                'type' => 'string',
+                            ),
                         ),
                     ),
                 ),
@@ -4379,6 +4666,22 @@ return array (
                             ),
                             'UseEbsOptimizedInstances' => array(
                                 'type' => 'boolean',
+                            ),
+                            'LifecycleEventConfiguration' => array(
+                                'type' => 'object',
+                                'properties' => array(
+                                    'Shutdown' => array(
+                                        'type' => 'object',
+                                        'properties' => array(
+                                            'ExecutionTimeout' => array(
+                                                'type' => 'numeric',
+                                            ),
+                                            'DelayUntilElbConnectionsDrained' => array(
+                                                'type' => 'boolean',
+                                            ),
+                                        ),
+                                    ),
+                                ),
                             ),
                         ),
                     ),
@@ -4640,6 +4943,23 @@ return array (
                 ),
             ),
         ),
+        'DescribeStackProvisioningParametersResult' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'AgentInstallerUrl' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+                'Parameters' => array(
+                    'type' => 'object',
+                    'location' => 'json',
+                    'additionalProperties' => array(
+                        'type' => 'string',
+                    ),
+                ),
+            ),
+        ),
         'DescribeStackSummaryResult' => array(
             'type' => 'object',
             'additionalProperties' => true,
@@ -4666,10 +4986,16 @@ return array (
                         'InstancesCount' => array(
                             'type' => 'object',
                             'properties' => array(
+                                'Assigning' => array(
+                                    'type' => 'numeric',
+                                ),
                                 'Booting' => array(
                                     'type' => 'numeric',
                                 ),
                                 'ConnectionLost' => array(
+                                    'type' => 'numeric',
+                                ),
+                                'Deregistering' => array(
                                     'type' => 'numeric',
                                 ),
                                 'Online' => array(
@@ -4679,6 +5005,12 @@ return array (
                                     'type' => 'numeric',
                                 ),
                                 'Rebooting' => array(
+                                    'type' => 'numeric',
+                                ),
+                                'Registered' => array(
+                                    'type' => 'numeric',
+                                ),
+                                'Registering' => array(
                                     'type' => 'numeric',
                                 ),
                                 'Requested' => array(
@@ -4706,6 +5038,9 @@ return array (
                                     'type' => 'numeric',
                                 ),
                                 'Terminating' => array(
+                                    'type' => 'numeric',
+                                ),
+                                'Unassigning' => array(
                                     'type' => 'numeric',
                                 ),
                             ),
@@ -5003,6 +5338,16 @@ return array (
             'additionalProperties' => true,
             'properties' => array(
                 'ElasticIp' => array(
+                    'type' => 'string',
+                    'location' => 'json',
+                ),
+            ),
+        ),
+        'RegisterInstanceResult' => array(
+            'type' => 'object',
+            'additionalProperties' => true,
+            'properties' => array(
+                'InstanceId' => array(
                     'type' => 'string',
                     'location' => 'json',
                 ),